refactor(remove-demo): find/update moved to tools, prettier only on dry run
