+++ src/commands/remove-demo.ts
@@ -1,6 +1,5 @@
 import { GluegunToolbox } from "gluegun"
 import * as pathlib from "path"
-import type { CommentType } from "../tools/demo"
 import { demo } from "../tools/demo"
 import { boolFlag } from "../tools/flag"
 import { p, warning } from "../tools/pretty"
@@ -23,7 +22,7 @@ module.exports = {
   description:
     "Remove demo code from generated boilerplate. Add --dry-run to see what would be removed.",
   run: async (toolbox: GluegunToolbox) => {
-    const { parameters, patching, filesystem } = toolbox
+    const { parameters, filesystem } = toolbox
 
     const CWD = process.cwd()
     const TARGET_DIR = parameters.first ?? CWD
@@ -32,60 +31,10 @@ module.exports = {
     p()
     p(`Removing demo code from '${TARGET_DIR}'${dryRun ? " (dry run)" : ""}`)
 
-    const filePaths = filesystem
-      .cwd(TARGET_DIR)
-      .find({
-        matching: MATCHING_GLOBS,
-        recursive: true,
-        files: true,
-        directories: false,
-      })
-      .map((path) => pathlib.join(TARGET_DIR, path))
+    const filePaths = demo.find(TARGET_DIR)
 
     // Go through every file path and handle the operation for each demo comment
-    const demoCommentResults = await Promise.allSettled(
-      filePaths.map(async (path) => {
-        const { exists, update } = patching
-        const { read } = filesystem
-        const {
-          REMOVE_CURRENT_LINE,
-          REMOVE_NEXT_LINE,
-          REMOVE_BLOCK_START,
-          REMOVE_BLOCK_END,
-          REMOVE_FILE,
-        } = demo.CommentType
-
-        const comments: CommentType[] = []
-
-        if (await exists(path, REMOVE_FILE)) {
-          if (!dryRun) filesystem.remove(path)
-          comments.push(REMOVE_FILE)
-          return { path, comments }
-        }
-
-        const operations = [
-          REMOVE_CURRENT_LINE,
-          REMOVE_NEXT_LINE,
-          REMOVE_BLOCK_START,
-          REMOVE_BLOCK_END,
-        ]
-        const shouldUpdate = RegExp(operations.join("|"), "g")
-
-        if (await exists(path, shouldUpdate)) {
-          const before = read(path)
-
-          operations.forEach((operation) => {
-            if (before.includes(operation)) {
-              comments.push(operation)
-            }
-          })
-
-          if (!dryRun) await update(path, demo.remove)
-        }
-
-        return { path, comments }
-      }),
-    )
+    const demoCommentResults = await demo.update({ filePaths, dryRun })
 
     // Handle the results of the demo comment operations
     demoCommentResults
